Add coroutine support
